edit.

+++ src/main/java/foamenbot/repositories/CategoryRepository.java
@@ -11,6 +11,6 @@ public interface CategoryRepository extends CrudRepository<Category, Long>{
 
     public List<Category> findAll();
     public Category findByCategoryName(String categoryName);
-    public Category findById(int id);
+    public Category findById(long id);
 
 }
