+++ src/main/java/foamenbot/services/CategoryService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface CategoryService {
 
     List<Category> findAll();
-    Category findById(int id);
+    Category findById(long id);
     Category findByCategoryName(String categoryName);
     void delete(Category category);
     void save(Category category);
