+++ src/main/java/foamenbot/services/IngredientService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IngredientService {
     List<Ingredient> findAll();
     Ingredient findByName(String name);
-    Ingredient findById(int id);
+    Ingredient findById(long id);
     void save(Ingredient ingredient);
     void delete(Ingredient ingredient);
 }
