+++ src/main/java/foamenbot/services/OrderProductService.java
@@ -6,7 +6,7 @@ import java.util.Set;
 
 public interface OrderProductService {
 
-    OrderProduct findById(int id);
+    OrderProduct findById(long id);
     void save(Set<OrderProduct> orderProducts);
     void save(OrderProduct orderProduct);
     void delete(Set<OrderProduct> orderProducts);
