+++ src/main/java/foamenbot/services/OrderService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface OrderService {
 
-    Order findById(int id);
+    Order findById(long id);
     void save(Order order);
     void delete(Order order);
     List<Order> findByStatus(String status);
